+++ sccfm_cli/commands/objects/__init__.py
@@ -1,0 +1,3 @@
+from sccfm_cli.commands.objects.command import ObjectsCommand
+
+__all__ = ["ObjectsCommand"]