+++ sccfm_cli/commands/objects/network/create/__init__.py
@@ -1,0 +1,3 @@
+from sccfm_cli.commands.objects.network.create.command import NetworkCreateCommand
+
+__all__ = ["NetworkCreateCommand"]