+++ sccfm_core/services/__init__.py
@@ -4,6 +4,7 @@ from sccfm_core.services.inventory import (
     AsaDiskFileService,
     InventoryService,
 )
+from sccfm_core.services.object_management import NetworkObjectService
 
 __all__ = [
     "AsaCommandLineService",
@@ -11,4 +12,5 @@ __all__ = [
     "HealthService",
     "HealthStatus",
     "InventoryService",
+    "NetworkObjectService",
 ]